Update test client
